:heavy_plus_sign:  Add typescript
