+++ src/pages/index.tsx
@@ -1,0 +1,19 @@
+import Head from 'next/head'
+
+export default function Home() {
+  return (
+    <div className="container">
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+
+      <main className="main">
+        <h1 className="title">
+          Welcome to <a href="https://nextjs.org">Next.js!</a>
+        </h1>
+      </main>
+    </div>
+  )
+}
